backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
pydantic[email]>=2.7.0
pydantic-settings>=2.2.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
alembic>=1.13.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
qrcode[pil]>=7.4.2 